week_3/pyproject.toml
[project]
name = "week3-csv-analytics"
version = "3.0.0"
description = "Containerized CSV Analytics with MCP Integration"
requires-python = ">=3.13"
dependencies = [
    # Core agent framework
    "openai-agents>=0.2.6",
    
    # OpenAI integration
    "openai>=1.99.1",
    
    # FastAPI and web server
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    
    # MCP server
    "fastmcp>=2.11.3",
    
    # Data processing and visualization
    "pandas>=2.3.1",
    "matplotlib>=3.10.5",
    
    # Phoenix tracing and observability
    "arize-phoenix>=11.21.0",
    "openinference-instrumentation-openai-agents>=1.2.0",
    "opentelemetry-api>=1.36.0",
    "opentelemetry-sdk>=1.36.0",
    
    # Utilities
    "python-dotenv>=1.1.1",
    "pydantic>=2.0.0",
]

[dependency-groups]
test = [
    "requests>=2.31.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]